refactor: 將各 schemas 與 router 分開
- 參考 [fastapi-best-practices]
(https://github.com/zhanymkanov/fastapi-best-practices#1-project-
structure-consistent--predictable) 所做的變更
