+++ src/api/routers/dining.py
@@ -1,64 +1,35 @@
-from enum import Enum
-from typing import List, Optional
-
 from fastapi import APIRouter, HTTPException, Path
-from pydantic import BaseModel, Field, HttpUrl
 
 from ..models.dining import Dining
-
-
-class DiningBuildingName(str, Enum):
-    小吃部 = "小吃部"
-    水木生活中心 = "水木生活中心"
-    風雲樓 = "風雲樓"
-    綜合教學大樓_南大校區 = "綜合教學大樓(南大校區)"
-    其他餐廳 = "其他餐廳"
-
-
-class DiningSceduleName(str, Enum):
-    weekday = "weekday"
-    saturday = "saturday"
-    sunday = "sunday"
-
-
-class DiningRestaurant(BaseModel):
-    area: str = Field(..., description="餐廳所在建築")
-    image: Optional[HttpUrl] = Field(..., description="餐廳圖片")
-    name: str = Field(..., description="餐廳名稱")
-    note: str = Field(..., description="餐廳備註")
-    phone: str = Field(..., description="餐廳電話")
-    schedule: dict = Field(..., description="餐廳營業時間")
-
-
-class DiningBuilding(BaseModel):
-    building: str = Field(..., description="建築名稱")
-    restaurants: list[DiningRestaurant] = Field(..., description="餐廳資料")
+from src.api import schemas
 
 
 router = APIRouter()
 dining = Dining()
 
 
-@router.get("/", response_model=list[DiningBuilding])
-def get_all_dining_data() -> list[DiningBuilding]:
+@router.get("/", response_model=list[schemas.dining.DiningBuilding])
+def get_all_dining_data() -> list[schemas.dining.DiningBuilding]:
     """
     取得所有餐廳資料。
     """
     return dining.get_dining_data()
 
 
-@router.get("/buildings", response_model=list[DiningBuildingName])
-def get_all_building_names() -> list[DiningBuildingName]:
+@router.get("/buildings", response_model=list[schemas.dining.DiningBuildingName])
+def get_all_building_names() -> list[schemas.dining.DiningBuildingName]:
     """
     取得所有建築名稱。
     """
     return dining.get_all_building_names()
 
 
-@router.get("/buildings/{building_name}", response_model=DiningBuilding)
+@router.get("/buildings/{building_name}", response_model=schemas.dining.DiningBuilding)
 def get_dining_data_in_buildings(
-    building_name: DiningBuildingName = Path(..., example="小吃部", description="建築名稱")
-) -> DiningBuilding:
+    building_name: schemas.dining.DiningBuildingName = Path(
+        ..., example="小吃部", description="建築名稱"
+    )
+) -> schemas.dining.DiningBuilding:
     """
     使用建築名稱取得指定建築的餐廳資料。
     """
@@ -73,7 +44,10 @@ def get_all_restaurant_names() -> list[str]:
     return dining.get_all_restaurant_names()
 
 
-@router.get("/restaurants/{restaurant_name}", response_model=list[DiningRestaurant])
+@router.get(
+    "/restaurants/{restaurant_name}",
+    response_model=list[schemas.dining.DiningRestaurant],
+)
 def get_dining_data_by_name(
     restaurant_name: str = Path(..., example="麥當勞", description="餐廳名稱")
 ):
@@ -83,10 +57,14 @@ def get_dining_data_by_name(
     return dining.query_by_restaurant_name(restaurant_name)
 
 
-@router.get("/scedules/{day_of_week}", response_model=list[DiningRestaurant])
+@router.get(
+    "/scedules/{day_of_week}", response_model=list[schemas.dining.DiningRestaurant]
+)
 def get_schedule_by_day_of_week(
-    day_of_week: DiningSceduleName = Path(..., example="saturday", description="營業日")
-) -> list[DiningRestaurant]:
+    day_of_week: schemas.dining.DiningSceduleName = Path(
+        ..., example="saturday", description="營業日"
+    )
+) -> list[schemas.dining.DiningRestaurant]:
     """
     取得所有該營業日的餐廳資訊。
     """
@@ -97,11 +75,12 @@ def get_schedule_by_day_of_week(
 
 
 @router.get(
-    "/searches/restaurants/{restaurant_name}", response_model=list[DiningRestaurant]
+    "/searches/restaurants/{restaurant_name}",
+    response_model=list[schemas.dining.DiningRestaurant],
 )
 def fuzzy_search_restaurant_by_name(
     restaurant_name: str = Path(..., example="麵", description="餐廳名稱")
-) -> List[DiningRestaurant]:
+) -> list[schemas.dining.DiningRestaurant]:
     """
     使用餐廳名稱模糊搜尋餐廳資料。
     """
