+++ src/api/routers/locations.py
@@ -1,35 +1,15 @@
 from uuid import UUID
 
 from fastapi import APIRouter, HTTPException, Path
-from pydantic import BaseModel, Field
 
 from ..models.locations import Location
-
-
-class LocationDetail(BaseModel):
-    name: str = Field(..., description="地點名稱")
-    name_en: str = Field(..., description="地點英文名稱")
-    latitude: str = Field(..., description="地點緯度")
-    longitude: str = Field(..., description="地點經度")
-
-
-class LocationData(BaseModel):
-    id: UUID = Field(..., description="地點 id")
-    data: LocationDetail = Field(..., description="地點資料")
-    create_time: str = Field(..., description="建立時間")
-    update_time: str = Field(..., description="更新時間")
-
-
-class LocationSearchData(BaseModel):
-    name: str = Field(..., description="要查詢的地點")
-    max_result: int = Field(10, description="最多回傳幾筆資料")
-
+from src.api import schemas
 
 router = APIRouter()
 location = Location()
 
 
-@router.get("/", response_model=list[LocationData])
+@router.get("/", response_model=list[schemas.locations.LocationData])
 def get_all_location():
     """
     取得所有地點資訊。
@@ -38,7 +18,7 @@ def get_all_location():
     return result
 
 
-@router.get("/{id}", response_model=LocationData)
+@router.get("/{id}", response_model=schemas.locations.LocationData)
 def get_location_by_id(
     id: UUID = Path(
         ..., example="b876aa09-40a8-427b-8bc7-1933978690e2", description="要查詢的地點 id"
@@ -53,7 +33,7 @@ def get_location_by_id(
     raise HTTPException(status_code=404, detail="Not found")
 
 
-@router.get("/searches/{name}", response_model=list[LocationData])
+@router.get("/searches/{name}", response_model=list[schemas.locations.LocationData])
 def search_location_by_get_method(
     name: str = Path(..., example="校門", description="要查詢的地點")
 ):
@@ -66,8 +46,8 @@ def search_location_by_get_method(
     return result
 
 
-@router.post("/searches", response_model=list[LocationData])
-def search_location_by_post_method(search_data: LocationSearchData):
+@router.post("/searches", response_model=list[schemas.locations.LocationData])
+def search_location_by_post_method(search_data: schemas.locations.LocationSearchData):
     """
     使用名稱模糊搜尋地點資訊。
     """
