+++ src/api/routers/resources/careers.py
@@ -1,11 +1,13 @@
 from fastapi import APIRouter
 
 from src.utils.scraper import rpage_scraper
+from src.api import schemas
+
 
 router = APIRouter()
 
 
-@router.get("/bulletin/recruitment")
+@router.get("/bulletin/recruitment", response_model=list[schemas.resources.RssData])
 async def get_bulletin_recruitment():
     """
     獲取清華公佈欄的徵才公告。
