+++ src/api/routers/rpage.py
@@ -4,18 +4,13 @@ from fastapi import APIRouter, Path
 from pydantic import BaseModel, Field, HttpUrl
 
 from src.utils.scraper import rpage_scraper
-
-
-class RpageData(BaseModel):
-    title: Optional[str] = Field(default=None, description="公告標題")
-    date: Optional[str] = Field(default=None, description="公告日期")
-    link: Optional[HttpUrl] = Field(default=None, description="公告網址")
+from src.api import schemas
 
 
 router = APIRouter()
 
 
-@router.get("/{full_path:path}", response_model=list[RpageData])
+@router.get("/{full_path:path}", response_model=list[schemas.resources.RssData])
 def get_rpage_data(
     full_path: HttpUrl = Path(
         ...,
