+++ src/api/schemas/__init__.py
@@ -1,0 +1,8 @@
+from . import resources
+from . import newsletter
+from . import locations
+from . import energy
+from . import dining
+from . import courses
+from . import contacts
+from . import buses
