+++ src/api/schemas/energy.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel, Field
+
+
+class ElectricityInfo(BaseModel):
+    name: str = Field(..., description="電力名稱")
+    data: int = Field(..., description="電力使用量")
+    capacity: int = Field(..., description="電力容量")
+    unit: str = Field(..., description="單位")
+    last_updated: str = Field(..., description="最後更新時間")
